refactor: AuthController에 계정 인증 API 통합
