+++ src/test/java/com/fmi/domain/auth/web/controller/AuthControllerTest.java
@@ -8,16 +8,23 @@ import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 import com.fmi.domain.Enum.Role;
+import com.fmi.domain.Enum.WithdrawalReason;
 import com.fmi.domain.auth.response.LoginResponse;
 import com.fmi.domain.auth.service.AuthService;
+import com.fmi.domain.auth.service.PasswordService;
 import com.fmi.domain.auth.service.TokenIssuer;
+import com.fmi.domain.auth.service.WithdrawalService;
+import com.fmi.domain.auth.web.dto.AccountDeleteRequest;
 import com.fmi.domain.auth.web.dto.LoginRequest;
+import com.fmi.domain.auth.web.dto.PasswordChangeRequest;
+import com.fmi.domain.auth.web.dto.PasswordVerifyRequest;
 import com.fmi.domain.user.data.User;
 import com.fmi.global.apiPayload.ApiResponse;
 import com.fmi.security.CookieFactory;
 import jakarta.servlet.http.Cookie;
 import java.time.Instant;
 import java.util.Date;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -31,7 +38,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.test.util.ReflectionTestUtils;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 
 @ExtendWith(MockitoExtension.class)
 class AuthControllerTest {
@@ -43,7 +54,16 @@ class AuthControllerTest {
     private TokenIssuer tokenIssuer;
 
     @Mock
+    private PasswordService passwordService;
+
+    @Mock
+    private WithdrawalService withdrawalService;
+
+    @Mock
     private CookieFactory cookieFactory;
+
+    @Mock
+    private UserDetails userDetails;
 
     @InjectMocks
     private AuthController authController;
@@ -203,6 +223,116 @@ class AuthControllerTest {
                 assertThat(response.getHeaders().get("Set-Cookie"))
                         .containsExactly("access_token=" + accessToken, "refresh_token=" + refreshToken);
                 assertThat(response.getBody().getResult()).isEqualTo(new LoginResponse(1L, true, true));
+            }
+        }
+    }
+
+    @Nested
+    @DisplayName("비밀번호 검증")
+    class VerifyPassword {
+
+        @Test
+        @DisplayName("기존 비밀번호 검증 경로를 유지한다")
+        void keepsPasswordVerifyPath() throws NoSuchMethodException {
+            PostMapping mapping = AuthController.class
+                    .getMethod("verifyPassword", UserDetails.class, PasswordVerifyRequest.class)
+                    .getAnnotation(PostMapping.class);
+
+            assertThat(mapping.value()).containsExactly("/users/me/password/verify");
+        }
+
+        @Test
+        @DisplayName("현재 비밀번호 검증을 비밀번호 유스케이스에 전달한다")
+        void verifiesCurrentPassword() {
+            // given
+            String email = "member@finditem.kr";
+            PasswordVerifyRequest request = new PasswordVerifyRequest();
+            request.setCurrentPassword("CurrentPassword1!");
+            when(userDetails.getUsername()).thenReturn(email);
+
+            // when
+            authController.verifyPassword(userDetails, request);
+
+            // then
+            verify(passwordService).verify(email, request);
+        }
+    }
+
+    @Nested
+    @DisplayName("비밀번호 변경")
+    class ChangePassword {
+
+        @Test
+        @DisplayName("기존 비밀번호 변경 경로를 유지한다")
+        void keepsPasswordChangePath() throws NoSuchMethodException {
+            PatchMapping mapping = AuthController.class
+                    .getMethod("changePassword", UserDetails.class, PasswordChangeRequest.class)
+                    .getAnnotation(PatchMapping.class);
+
+            assertThat(mapping.value()).containsExactly("/users/me/password");
+        }
+
+        @Test
+        @DisplayName("요청 값을 해체해 비밀번호 유스케이스에 전달한다")
+        void changesPassword() {
+            // given
+            String email = "member@finditem.kr";
+            PasswordChangeRequest request = new PasswordChangeRequest();
+            request.setNewPassword("NewPassword1!");
+            request.setNewPasswordConfirm("NewPassword1!");
+            when(userDetails.getUsername()).thenReturn(email);
+
+            // when
+            authController.changePassword(userDetails, request);
+
+            // then
+            verify(passwordService).change(email, "NewPassword1!", "NewPassword1!");
+        }
+    }
+
+    @Nested
+    @DisplayName("회원 탈퇴")
+    class DeleteAccount {
+
+        @Test
+        @DisplayName("기존 회원 탈퇴 경로를 유지한다")
+        void keepsAccountDeletePath() throws NoSuchMethodException {
+            DeleteMapping mapping = AuthController.class
+                    .getMethod(
+                            "deleteAccount",
+                            UserDetails.class,
+                            AccountDeleteRequest.class,
+                            jakarta.servlet.http.HttpServletRequest.class)
+                    .getAnnotation(DeleteMapping.class);
+
+            assertThat(mapping.value()).containsExactly("/users/me");
+        }
+
+        @Nested
+        @DisplayName("탈퇴 요청이 유효하면")
+        class WithValidRequest {
+
+            @Test
+            @DisplayName("탈퇴를 처리하고 액세스 쿠키와 리프레시 쿠키를 순서대로 만료한다")
+            void deletesAccountAndExpiresCookiesInOrder() {
+                // given
+                String email = "member@finditem.kr";
+                AccountDeleteRequest request = new AccountDeleteRequest();
+                request.setReasons(List.of(WithdrawalReason.NOT_USING));
+                MockHttpServletRequest httpRequest = new MockHttpServletRequest();
+                when(userDetails.getUsername()).thenReturn(email);
+                when(cookieFactory.expire(eq(httpRequest), eq("access_token")))
+                        .thenReturn(ResponseCookie.from("access_token", "").build());
+                when(cookieFactory.expire(eq(httpRequest), eq("refresh_token")))
+                        .thenReturn(ResponseCookie.from("refresh_token", "").build());
+
+                // when
+                ResponseEntity<ApiResponse<Void>> response =
+                        authController.deleteAccount(userDetails, request, httpRequest);
+
+                // then
+                verify(withdrawalService).delete(email, request);
+                assertThat(response.getHeaders().get("Set-Cookie")).containsExactly("access_token=", "refresh_token=");
             }
         }
     }
